PMM-6542 styles cleanup
