+++ pmm-app/src/pmm-add-instance-app-panel/panel.scss
@@ -1,12 +1,8 @@
-.add-instance-panel {
-  display: flex;
-  flex-direction: column;
-  align-content: center;
-  align-items: center;
+.padding {
   padding-top: 20px;
 }
 
-.add-instance-navigation-link {
+.flex-column {
   display: flex;
   flex-direction: column;
   align-content: center;
